memory

+++ src/api/makronexusAI/langchain/index.js
@@ -2,12 +2,12 @@
 import express from 'express';
 import {DirectoryLoader} from "langchain/document_loaders/fs/directory"
 import { OpenAI } from 'langchain/llms/openai';
-import { RetrievalQAChain, loadQAStuffChain } from 'langchain/chains';
+import { ConversationalRetrievalQAChain, RetrievalQAChain, loadQAStuffChain } from 'langchain/chains';
 import { OpenAIEmbeddings } from 'langchain/embeddings/openai';
 import { FaissStore } from 'langchain/vectorstores/faiss';
 import { QdrantVectorStore } from 'langchain/vectorstores/qdrant';
 import { TextLoader } from 'langchain/document_loaders/fs/text';
-import { CharacterTextSplitter } from 'langchain/text_splitter';
+import { CharacterTextSplitter, RecursiveCharacterTextSplitter } from 'langchain/text_splitter';
 import { PDFLoader } from 'langchain/document_loaders/fs/pdf';
 import { CSVLoader } from 'langchain/document_loaders/fs/csv';
 import { DocxLoader } from 'langchain/document_loaders/fs/docx';
@@ -110,37 +110,56 @@ router.post('/:user_id/:userAISettings_id/files/save',upload.single('file'), asy
      );
      const tokens=encoder.encode(JSON.stringify(docs));
      const tokenCount=tokens.length;
-     const ratePerThousandTokens=0.002
+     const ratePerThousandTokens=0.004
      const cost=(tokenCount/1000)*ratePerThousandTokens;
      encoder.free();
      return cost
    }
-   const splitter = new CharacterTextSplitter({
-         chuckSize: 1000,
-         chunkOverlap: 0,
+   const normalizeDocuments=(docs)=>{
+    return docs.map((doc)=>{
+      if(typeof doc.pageContent==="string"){
+        return doc.pageContent;
+      }else if (Array.isArray(doc.pageContent)){
+        return doc.pageContent.join("\n")
+      }
+    });
+   }
+   const cost=await calculateCost()
+   console.log("calculated cost:",cost)
+   if(cost<1){
+    const splitter = new RecursiveCharacterTextSplitter({
+      chuckSize: 1000,
+      chunkOverlap: 0,
+    });
+    const normalizedDocs=normalizeDocuments(docs)
+    const documents = await splitter.createDocuments(normalizedDocs);
+        const embeddings = new OpenAIEmbeddings();
+    const vectorstore = await QdrantVectorStore.fromDocuments(documents, embeddings,
+      {
+        url: process.env.QDRANT_URL,
+         collectionName: collectionName,
+         apiKey:process.env.QDRANT_DB_KEY
+      });
+      console.log(vectorstore,"VECTORSTORE")
+    if (vectorstore) {
+      try {
+        // Delete all files in the "uploads" folder
+       const uploadFolderPath = getFilePath('../../../../uploads');
+       fs.readdirSync(uploadFolderPath).forEach((file) => {
+         const filePath = path.join(uploadFolderPath, file);
+         fs.unlinkSync(filePath);
        });
-       const documents = await splitter.splitDocuments(docs);
-           const embeddings = new OpenAIEmbeddings();
-       const vectorstore = await QdrantVectorStore.fromDocuments(documents, embeddings,
-         {
-           url: process.env.QDRANT_URL,
-            collectionName: collectionName,
-            apiKey:process.env.QDRANT_DB_KEY
-         });
-         console.log(vectorstore,"VECTORSTORE")
-       if (vectorstore) {
-         try {
-           // Delete all files in the "uploads" folder
-          const uploadFolderPath = getFilePath('../../../../uploads');
-          fs.readdirSync(uploadFolderPath).forEach((file) => {
-            const filePath = path.join(uploadFolderPath, file);
-            fs.unlinkSync(filePath);
-          });
-           res.status(201).json({ message: 'File saved successfully',file:file });
-          } catch (error) {
-            console.error('Error deleting file:', error);
-          }
-        }
+        res.status(201).json({ message: 'File saved successfully',file:file,cost:"USD" +cost });
+       } catch (error) {
+         console.error('Error deleting file:', error);
+       }
+     }
+
+   }else{
+    console.log("Cost exceeds 1 USD")
+    res.status(401).json({ message: 'Cost of embedding exceeds  $1. Skipping embeddings '});
+   }
+
        
        }
   } catch (error) {
@@ -183,11 +202,31 @@ router.post('/:user_id/:dataset_id/chats/:chat_id/query', async (req, res) => {
     });
 
     const model = new OpenAI({ temperature: temperature||0.3, model: "gpt-3.5-turbo", });
-    const chain = new RetrievalQAChain({
-      combineDocumentsChain: loadQAStuffChain(model),
-      retriever: vectorStore.asRetriever(),
-      returnSourceDocuments: true,
-    });
+    const chatPrompt = ChatPromptTemplate.fromPromptMessages([
+      SystemMessagePromptTemplate.fromTemplate(
+        `You are a Socratic tutor. Use the following principles in responding to students:\n
+        - Ask thought-provoking, open-ended questions that challenge students' preconceptions and encourage them to engage in deeper reflection and critical thinking.\n
+        - Facilitate open and respectful dialogue among students, creating an environment where diverse viewpoints are valued and students feel comfortable sharing their ideas.\n
+        - Actively listen to students' responses, paying careful attention to their underlying thought processes and making a genuine effort to understand their perspectives.\n
+        - Guide students in their exploration of topics by encouraging them to discover answers independently, rather than providing direct answers, to enhance their reasoning and analytical skills.\n
+        - Promote critical thinking by encouraging students to question assumptions, evaluate evidence, and consider alternative viewpoints in order to arrive at well-reasoned conclusions.\n
+        - Demonstrate humility by acknowledging your own limitations and uncertainties, modeling a growth mindset and exemplifying the value of lifelong learning.`
+      ),
+      new MessagesPlaceholder('chat_history'),
+      HumanMessagePromptTemplate.fromTemplate('{query}'),
+    ]);
+    const chain = new ConversationalRetrievalQAChain.fromLLM(
+      model,
+      {prompt: chatPrompt},
+      { returnSourceDocuments: true},
+      vectorStore.asRetriever(),
+      {
+        memory:new BufferMemory({
+          memoryKey:"chat_history",
+          returnMessages: true,
+        })
+      }
+    );
 
     const result = await chain.call({
       query: question,
@@ -225,6 +264,82 @@ router.post('/:user_id/:dataset_id/chats/:chat_id/query', async (req, res) => {
     res.status(500).json({ error: 'An error occurred while querying the file' });
   }
 });
+// router.post('/:user_id/:dataset_id/chats/:chat_id/query', async (req, res) => {
+//   // const collectionName = "My Cover letter";
+//   const { question,collectionName,temperature,personality, } = req.body;
+//   const { message,chat_id, dataset_id, user_id } = req.params;
+
+//   try {
+//     // Fetch user, chat, and dataset in parallel
+//     const [user, chat, dataset] = await Promise.all([
+//       UserModel.findByPk(user_id),
+//       aiChatModel.findByPk(chat_id),
+//       UserAISettingsModel.findByPk(dataset_id)
+
+//     ]);
+
+//     // Check for missing user or chat
+//     if (!user) {
+//       return res.status(404).json({ error: "User not found" });
+//     }
+
+//     if (!chat) {
+//       return res.status(404).json({ error: "Chat not found" });
+//     }
+//     if (!dataset) {
+//       return res.status(404).json({ error: "Dataset not found" });
+//     }
+     
+//     const embeddings = new OpenAIEmbeddings();
+//     const vectorStore = await QdrantVectorStore.fromExistingCollection(embeddings, {
+//       url: process.env.QDRANT_URL,
+//       collectionName: collectionName,
+//       apiKey: process.env.QDRANT_DB_KEY,
+//     });
+
+//     const model = new OpenAI({ temperature: temperature||0.3, model: "gpt-3.5-turbo", });
+//     const chain = new RetrievalQAChain({
+//       combineDocumentsChain: loadQAStuffChain(model),
+//       retriever: vectorStore.asRetriever(),
+//       returnSourceDocuments: true,
+//     });
+
+//     const result = await chain.call({
+//       query: question,
+//     });
+
+//     // Create a new DatasetChatModel instance for the user's input
+//     const newDatasetChat = await DatasetChatModel.create({
+//       type: "text",
+//       message: question,
+//       from: "user",
+//       model: "gpt-3.5-turbo",
+//       user_id: user_id,
+//       chat_id: chat_id,
+//       dataset_id: dataset_id,
+//     });
+ 
+    
+//     // Associate the user's input with the user
+//     await newDatasetChat.setUser(user);
+
+//     // Create a new DatasetChatModel instance for the AI response
+//     const newResponseDatasetChat = await DatasetChatModel.create({
+//       type: "text",
+//       message: result.text,
+//       model: "gpt-3.5-turbo",
+//       user_id: process.env.MAKRONEXA_ID, //cloud
+//       chat_id: chat_id,
+//       dataset_id: dataset_id,
+//     });
+
+//     console.log(newResponseDatasetChat, "newResponseDatasetChat");
+//     res.json({ message: result.text });
+//   } catch (error) {
+//     console.error(error);
+//     res.status(500).json({ error: 'An error occurred while querying the file' });
+//   }
+// });
 
 
 router.post('/:user_id/:dataset_id/chats/:chat_id/chat', async (req, res) => {
@@ -253,12 +368,7 @@ router.post('/:user_id/:dataset_id/chats/:chat_id/chat', async (req, res) => {
     const response = await chain.call({
       input: req.body.question,
     });
-    // const response = await chain.call({
-    //   input: 'how old is Mr Gatsby?',
-    // });
-    // const response2 = await chain.call({
-    //   input: 'where does he live?',
-    // });
+    
 
     res.json( response );
   } catch (error) {
